requirements.txt
requests
beautifulsoup4
pandas
tqdm
python-dotenv